Refactor processing images
Deleted duplicated code
Added mutex to images downloading

+++ common/core-files/src/androidMain/kotlin/com/security/chat/multiplatform/common/core/files/FileManager.android.kt
@@ -80,7 +80,7 @@ internal class FileManagerAndroid(
     ): String {
         return withContext(dispatcherProvider.IO) {
             val destinationFile = File(
-                getOrCreateDirectory(directoryName),
+                getOrCreateCacheDirectory(directoryName),
                 UUID.randomUUID().toString(),
             )
 
@@ -95,9 +95,9 @@ internal class FileManagerAndroid(
         }
     }
 
-    override suspend fun getDirectoryPath(name: String): String {
+    override suspend fun getCacheDirectoryPath(name: String): String {
         return withContext(dispatcherProvider.IO) {
-            getOrCreateDirectory(name).absolutePath
+            getOrCreateCacheDirectory(name).absolutePath
         }
     }
 
@@ -116,6 +116,9 @@ internal class FileManagerAndroid(
             val source = File(sourcePath)
             val destination = File(destinationPath)
             destination.parentFile?.mkdirs()
+
+            if (source.renameTo(destination)) return@withContext
+
             source.copyTo(destination, overwrite = true)
             source.delete()
         }
@@ -127,17 +130,25 @@ internal class FileManagerAndroid(
         }
     }
 
-    override suspend fun clearDirectory(name: String) {
-        withContext(dispatcherProvider.IO) {
-            val directory = cacheDirectory.resolve(name)
-            directory.deleteRecursively()
-            check(directory.mkdirs()) {
-                "Cannot create cache directory: ${directory.absolutePath}"
-            }
+    override suspend fun fileExists(path: String): Boolean {
+        return withContext(dispatcherProvider.IO) {
+            File(path).isFile
         }
     }
 
-    private fun getOrCreateDirectory(name: String): File {
+    override suspend fun clearCacheDirectory(name: String) {
+        withContext(dispatcherProvider.IO) {
+            cacheDirectory.resolve(name).deleteRecursively()
+        }
+    }
+
+    override suspend fun clearDataDirectory(name: String) {
+        withContext(dispatcherProvider.IO) {
+            dataDirectory.resolve(name).deleteRecursively()
+        }
+    }
+
+    private fun getOrCreateCacheDirectory(name: String): File {
         val directory = cacheDirectory.resolve(name)
         check(directory.isDirectory || directory.mkdirs()) {
             "Cannot create cache directory: ${directory.absolutePath}"
